requirements/dev_requirements.txt
# Put requeriments just for dev env



-r requirements.txt
coverage==6.3.2
lagom==1.7.0
mimesis==5.3.0
moto==3.1.5
pre-commit==2.18.1
pylint==2.13.7
pytest==7.1.1
pytest-cov==3.0.0
python-dotenv==0.20.0
